Move test to their correct location
